Restructure library and examples
