+++ .scalafmt.conf
@@ -1,0 +1,22 @@
+style = defaultWithAlign
+maxColumn = 120
+
+continuationIndent.callSite = 2
+
+newlines {
+  sometimesBeforeColonInMethodReturnType = false
+}
+
+align {
+  arrowEnumeratorGenerator = false
+  ifWhileOpenParen = false
+  openParenCallSite = false
+  openParenDefnSite = false
+}
+
+docstrings = JavaDoc
+
+rewrite {
+  rules = [SortImports, RedundantBraces, RedundantParens]
+  redundantBraces.maxLines = 1
+}
