+++ build.sbt
@@ -28,6 +28,7 @@ lazy val V = new {
   val json4sVersion = "3.6.4"
   val mockitoVersion = "1.10.19"
   val scalazZIOVersion = "0.6.3"
+  val scalazCoreVersion = "7.2.27"
 }
 
 val noPublishSettings = Seq(
@@ -49,8 +50,6 @@ val buildSettings = Seq(
 
 val commonDependencies = Seq(
   libraryDependencies ++= Seq(
-    "org.typelevel" %% "cats-core" % V.catsVersion,
-    "org.typelevel" %% "cats-effect" % V.catsEffectVersion,
     "org.scalatest" %% "scalatest" % V.scalaTestVersion % Test,
     "org.mockito" % "mockito-all" % V.mockitoVersion % Test
   )
@@ -75,7 +74,7 @@ lazy val logger4s = project
 lazy val core = crossProject(JVMPlatform)
   .crossType(CrossType.Full)
   .in(file("core"))
-  .settings(moduleName := "logger4s")
+  .settings(moduleName := "logger4s-core")
   .settings(buildSettings)
   .settings(commonDependencies)
   .jvmSettings(libraryDependencies ++= Seq(
@@ -83,6 +82,21 @@ lazy val core = crossProject(JVMPlatform)
   ))
 
 lazy val coreJVM = core.jvm
+
+lazy val cats = crossProject(JVMPlatform)
+  .crossType(CrossType.Pure)
+  .in(file("logger4s-cats"))
+  .settings(moduleName := "logger4s-cats")
+  .settings(buildSettings)
+  .settings(commonDependencies)
+  .settings(compilerPlugins)
+  .settings(libraryDependencies ++= Seq(
+      "org.typelevel" %% "cats-core" % V.catsVersion,
+      "org.typelevel" %% "cats-effect" % V.catsEffectVersion,
+  ))
+  .dependsOn(core)
+
+lazy val catsJVM = cats.jvm
 
 lazy val scalaz = crossProject(JVMPlatform)
   .crossType(CrossType.Pure)
@@ -92,6 +106,7 @@ lazy val scalaz = crossProject(JVMPlatform)
   .settings(commonDependencies)
   .settings(compilerPlugins)
   .settings(libraryDependencies ++= Seq(
+    "org.scalaz" %% "scalaz-core" % V.scalazCoreVersion,
     "org.scalaz" %% "scalaz-zio" % V.scalazZIOVersion
   ))
   .dependsOn(core)
@@ -99,10 +114,10 @@ lazy val scalaz = crossProject(JVMPlatform)
 lazy val scalazJVM = scalaz.jvm
 
 lazy val example = project
-  .in(file("example"))
+  .in(file("example-cats"))
   .settings(buildSettings)
   .settings(noPublishSettings)
-  .dependsOn(coreJVM)
+  .dependsOn(coreJVM, catsJVM)
   .settings(
     libraryDependencies ++= Seq(
       "org.json4s" %% "json4s-native" % V.json4sVersion
