+++ example-cats/src/main/resources/logback.xml
